+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Fixed the previously inoperational methods `append`, `append_left`, `extend`, and `extend_left`
   of the class `CHATReader` through the upstream PyLangAcq package.
 - Retrained the part-of-speech tagger, after the minor character fix from v3.2.3.
+- Raised `NotImplementedError` for the method `ipsyn` of `CHATReader`,
+  since the upstream method works only for English.
 
 ## [3.2.3] - 2021-04-12